Acoplada/requirements.txt
Flask==3.0.3
pydantic==2.9.2
psycopg2-binary==2.9.9
botocore==1.34.34
